Remove duplicated module
